% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-geom-signif.R
\name{geom_signif}
\alias{geom_signif}
\title{geom_signif exported from ggsignif package}
\description{
See \code{ggsignif::\link[ggsignif:stat_signif]{geom_signif}} for details.
}
\examples{
ggplot2::ggplot(iris, ggplot2::aes(Species, Sepal.Length)) +
  ggplot2::geom_boxplot() +
  geom_signif(
    comparisons = list(c("versicolor", "virginica")),
    map_signif_level = TRUE
  )
}
\keyword{internal}
