% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpaired.R
\name{ggpaired}
\alias{ggpaired}
\title{Plot Paired Data}
\usage{
ggpaired(data, cond1, cond2, color = "black", fill = "white",
  palette = NULL, size = NULL, width = 0.7, line.color = "black",
  title = NULL, xlab = "Condition", ylab = "Value",
  ggtheme = theme_pubr(), ...)
}
\arguments{
\item{data}{a data frame}

\item{cond1}{variable name corresponding to the first condition.}

\item{cond2}{variable name corresponding to the second condition.}

\item{color}{points and boxs plot colors. To color by conditions, use color = "condition".}

\item{fill}{box plot fill color. To change fill color by conditions, use fill = "condition".}

\item{palette}{the color palette to be used for coloring or filling by groups.
Allowed values include "grey" for grey color palettes; brewer palettes e.g.
"RdBu", "Blues", ...; or custom color palette e.g. c("blue", "red"); and
scientific journal palettes from ggsci R package, e.g.: "npg", "aaas",
"lancet", "jco", "ucscgb", "uchicago", "simpsons" and "rickandmorty".}

\item{size}{line and point size.}

\item{width}{box plot width.}

\item{line.color}{line color.}

\item{title}{plot main title.}

\item{xlab}{character vector specifying x axis labels. Use xlab = FALSE to
hide xlab.}

\item{ylab}{character vector specifying y axis labels. Use ylab = FALSE to
hide ylab.}

\item{ggtheme}{function, ggplot2 theme name. Default value is theme_pubr().
Allowed values include ggplot2 official themes: theme_gray(), theme_bw(),
theme_minimal(), theme_classic(), theme_void(), ....}

\item{...}{other arguments to be passed to be passed to \link{ggpar}().}
}
\description{
Plot paired data.
}
\examples{

before <-c(200.1, 190.9, 192.7, 213, 241.4, 196.9, 172.2, 185.5, 205.2, 193.7)
after <-c(392.9, 393.2, 345.1, 393, 434, 427.9, 422, 383.9, 392.3, 352.2)

d <- data.frame(before = before, after = after)
ggpaired(d, cond1 = "before", cond2 = "after",
    fill = "condition", size = 0.6)


}
