% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{ggpubr-common-params}
\alias{ggpubr-common-params}
\title{Common parameters for ggpubr stat and geom functions}
\arguments{
\item{inherit.aes}{If \code{FALSE},
overrides the default aesthetics, rather than combining with them. This is
most useful for helper functions that define both data and aesthetics and
shouldn't inherit behaviour from the default plot specification. Set to
\code{TRUE} to inherit aesthetics from the parent ggplot layer. The
function usage states the default for each layer.}
}
\description{
Common parameters for ggpubr stat and geom functions
}
\keyword{internal}
