% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_comparison.R
\name{mc_test}
\alias{mc_test}
\title{Comparison of Means}
\usage{
mc_test(formula, data, method = "t.test", group.by = NULL, ...)
}
\arguments{
\item{formula}{a formula of the form \code{x ~ group} where \code{x} is a numeric variable
giving the data values and \code{group} is a factor with one or multiple levels giving the
corresponding groups.}

\item{data}{a data.frame containing the variables in the formula.}

\item{method}{the type of test}

\item{group.by}{a character vector containing the name of grouping
variables.}

\item{...}{Other arguments to be passed to the test function.}
}
\description{
Performs one or multiple mean comparisons.
}
\examples{
# Load data
#::::::::::::::::::::::::::::
data("ToothGrowth")
df <- ToothGrowth

# Compare supp levels after grouping the data by "dose"
#:::::::::::::::::::::::::::::::::::
mc_test(len ~ supp, df, group.by = "dose")

}
