% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mutate}
\alias{group_by}
\alias{get_summary_stats}
\alias{get_test_label}
\alias{get_pwc_label}
\alias{create_test_label}
\alias{add_cld}
\alias{add_xy_position}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}}

  \item{rstatix}{\code{\link[rstatix]{add_cld}}, \code{\link[rstatix:get_pvalue_position]{add_xy_position}}, \code{\link[rstatix:get_test_label]{create_test_label}}, \code{\link[rstatix:get_test_label]{get_pwc_label}}, \code{\link[rstatix]{get_summary_stats}}, \code{\link[rstatix]{get_test_label}}}
}}

