% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_line_types.R
\name{show_line_types}
\alias{show_line_types}
\title{Line types available in R}
\usage{
show_line_types()
}
\value{
a ggplot.
}
\description{
Show line types available in R.
}
\examples{
show_line_types()+
 theme_minimal()
}
\seealso{
\code{\link{ggpar}} and \code{\link{ggline}}.
}
