% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_point_shapes.R
\name{show_point_shapes}
\alias{show_point_shapes}
\title{Point shapes available in R}
\usage{
show_point_shapes()
}
\value{
a ggplot.
}
\description{
Show point shapes available in R.
}
\examples{
show_point_shapes()+
 theme_minimal()
}
\seealso{
\code{\link{ggpar}} and \code{\link{ggline}}.
}

