% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_transparent.R
\name{theme_transparent}
\alias{theme_transparent}
\title{Create a ggplot with Transparent Background}
\usage{
theme_transparent(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Create a ggplot with transparent background.
}
\examples{
# Create a scatter plot
sp <- ggscatter(iris, x = "Sepal.Length", y = "Sepal.Width",
               color = "Species", palette = "jco",
               size = 3, alpha = 0.6)
sp

# Transparent theme
sp + theme_transparent()

}
\seealso{
\code{\link{theme_pubr}}
}
