% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_figure.R
\name{annotate_figure}
\alias{annotate_figure}
\title{Annotate Arranged Figure}
\usage{
annotate_figure(p, top = NULL, bottom = NULL, left = NULL, right = NULL,
  fig.lab = NULL, fig.lab.pos = c("top.left", "top", "top.right",
  "bottom.left", "bottom", "bottom.right"), fig.lab.size, fig.lab.face)
}
\arguments{
\item{p}{(arranged) ggplots.}

\item{top, bottom, left, right}{optional string, or grob.}

\item{fig.lab}{figure label (e.g.: "Figure 1").}

\item{fig.lab.pos}{position of the figure label, can be one of "top.left",
"top", "top.right", "bottom.left", "bottom", "bottom.right". Default is
"top.left".}

\item{fig.lab.size}{optional size of the figure label.}

\item{fig.lab.face}{optional font face of the figure label. Allowed values
include: "plain", "bold", "italic", "bold.italic".}
}
\description{
Annotate figures including: i) ggplots, ii) arranged ggplots from
 \code{\link{ggarrange}()}, \code{\link[gridExtra]{grid.arrange}()} and
 \code{\link[cowplot]{plot_grid}()}.
}
\examples{
data("ToothGrowth")
df <- ToothGrowth
df$dose <- as.factor(df$dose)

# Create some plots
# ::::::::::::::::::::::::::::::::::::::::::::::::::
# Box plot
bxp <- ggboxplot(df, x = "dose", y = "len",
    color = "dose", palette = "jco")
# Dot plot
dp <- ggdotplot(df, x = "dose", y = "len",
    color = "dose", palette = "jco")
# Density plot
dens <- ggdensity(df, x = "len", fill = "dose", palette = "jco")

# Arrange and annotate
# ::::::::::::::::::::::::::::::::::::::::::::::::::
library(grid)
gp = grid::gpar(col = "black", fontface = "bold")
res <- ggarrange(bxp, dp, dens, ncol = 2, nrow = 2)
annotate_figure(res,
               top = textGrob("Visualizing Tooth Growth by Vitamin C Dose", gp = gp),
               bottom = "Data source: ToothGrowth data set",
               left = "Arranged plot using ggpubr",
               right = "I'm done",
               fig.lab = "Figure 1", fig.lab.face = "bold")


}
\seealso{
\code{\link{ggarrange}()}
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
