% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgcolor.R
\name{bgcolor}
\alias{bgcolor}
\title{Change ggplot Panel Background Color}
\usage{
bgcolor(color)
}
\arguments{
\item{color}{background color.}
}
\description{
Change ggplot panel background color.
}
\examples{
# Load data
data("ToothGrowth")
df <- ToothGrowth

# Basic plot
p <- ggboxplot(df, x = "dose", y = "len")
p

# Change panel background color
p +
  bgcolor("#BFD5E3")+
  border("#BFD5E3")
}
\seealso{
\link{border}().
}
