% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.R
\name{border}
\alias{border}
\title{Set ggplot Panel Border Line}
\usage{
border(color = "black", size = 0.8, linetype = NULL)
}
\arguments{
\item{color}{border line color.}

\item{size}{numeric value specifying border line size.}

\item{linetype}{line type. An integer (0:8), a name (blank, solid, dashed,
dotted, dotdash, longdash, twodash). Sess \code{\link{show_line_types}}.}
}
\description{
Change or set ggplot panel border.
}
\examples{
# Load data
data("ToothGrowth")
df <- ToothGrowth

# Basic plot
p <- ggboxplot(df, x = "dose", y = "len")
p

# Add border
p + border()
}
