% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggarrange.R
\name{ggarrange}
\alias{ggarrange}
\title{Arrange Multiple ggplots}
\usage{
ggarrange(..., plotlist = NULL, ncol = NULL, nrow = NULL, labels = NULL)
}
\arguments{
\item{...}{List of plots to be arranged into the grid. The plots can be either ggplot2 plot objects
or arbitrary gtables.}

\item{plotlist}{(optional) List of plots to display. Alternatively, the plots can be provided
individually as the first n arguments of the function plot_grid (see examples).}

\item{ncol}{(optional) Number of columns in the plot grid.}

\item{nrow}{(optional) Number of rows in the plot grid.}

\item{labels}{(optional) List of labels to be added to the plots. You can also set \code{labels="AUTO"} to
auto-generate upper-case labels or \code{labels="auto"} to auto-generate lower-case labels.}
}
\value{
a ggplot
}
\description{
Arrange multiple ggplots on the same page.
}
\examples{
data("ToothGrowth")
df <- ToothGrowth
df$dose <- as.factor(df$dose)

# Box plot
bxp <- ggboxplot(df, x = "dose", y = "len",
    color = "dose", palette = "jco")
# Dot plot
dp <- ggdotplot(df, x = "dose", y = "len",
    color = "dose", palette = "jco")
# Density plot
dens <- ggdensity(df, x = "len", fill = "dose", palette = "jco")

ggarrange(bxp, dp, dens, ncol = 2)

}
\author{
Alboukadel Kassambara <alboukadel.kassambara@gmail.com>
}
