% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboxplot.R
\name{ggboxplot}
\alias{ggboxplot}
\title{Box plot}
\usage{
ggboxplot(data, x, y, color = "black", fill = "white", palette = NULL,
  linetype = "solid", size = 1, width = 1, notch = FALSE,
  select = NULL, order = NULL, add = c("none", "dotplot", "jitter"),
  add.params = list(), ...)
}
\arguments{
\item{data}{a data frame}

\item{x, y}{x and y variables for drawing.}

\item{color, fill}{outline and fill colors.}

\item{palette}{the color palette to be used for coloring or filling by
groups. Allowed values include "grey" for grey color palettes; RColorBrewer
palettes e.g. "RdBu", "Blues", ...; or a character vector containing custom
color palette e.g. c("blue", "red").}

\item{linetype}{line types.}

\item{size}{line size.}

\item{width}{box plot width.}

\item{notch}{if \code{FALSE} (default) make a standard box plot. If
\code{TRUE}, make a notched box plot. Notches are used to compare groups;
if the notches of two boxes do not overlap, this suggests that the medians
are significantly different.}

\item{select}{character vector specifying which items to display.}

\item{order}{character vector specifying the order of items.}

\item{add}{character vector for adding another plot element. Allowed values
are one of c("none", "dotplot", "jitter).}

\item{add.params}{parameters (color, shape, size, fill, jitter) for the
argument 'add'; e.g.: add.params = list(color = "red").}

\item{...}{other arguments to be passed to geom_boxplot including linetype,
size, etc. (See ?geom_boxplot).}
}
\description{
Box plot
}
\details{
The plot can be easily customized using the function ggpar(). Read
  ?ggpar for changing: \itemize{ \item main title and axis labels: main,
  xlab, ylab \item axis limits: xlim, ylim (e.g.: ylim = c(0, 30)) \item axis
  scales: xscale, yscale (e.g.: yscale = "log2") \item color palettes:
  palette = "Dark2" or palette = c("gray", "blue", "red") \item legend title,
  labels and position: legend = "right" \item plot orientation : orientation
  = c("vertical", "horizontal", "reverse") }
}
\examples{
# Load data
data("ToothGrowth")
df <- ToothGrowth

# Basic plot
# +++++++++++++++++++++++++++
ggboxplot(df, x = "dose", y = "len")

# Change width
ggboxplot(df, "dose", "len", width = 0.2)

# Change orientation: horizontal
ggboxplot(df, "dose", "len", orientation = "horizontal")

# Notched box plot
ggboxplot(df, x = "dose", y = "len",
   notch = TRUE)

# Add dot
# ++++++++++++++++++++++++++
ggboxplot(df, x = "dose", y = "len",
   add = "dotplot")

# Add jitter
ggboxplot(df, x = "dose", y = "len",
   add = "jitter")

# Change jitter shape by groups: "dose"
ggboxplot(df, x = "dose", y = "len",
  add = "jitter", shape = "dose")


# Select and order items
# ++++++++++++++++++++++++++++++

# Select which items to display: "0.5" and "2"
ggboxplot(df, "dose", "len",
   select = c("0.5", "2"))

# Change the default order of items
ggboxplot(df, "dose", "len",
   order = c("2", "1", "0.5"))


# Change colors
# +++++++++++++++++++++++++++
# Change outline and fill colors
# Change line size

 ggboxplot(df, "dose", "len",
   color = "#BFD5E3", fill = "#6D9EC1", size = 1.2)

# Change outline colors by groups: dose
 ggboxplot(df, "dose", "len", color = "dose")

# Change fill color by groups: dose
 ggboxplot(df, "dose", "len", fill = "dose")

# Use custom color palette
 ggboxplot(df, "dose", "len", color = "dose",
  palette = c('#999999','#E69F00','#56B4E9') )

# Use brewer palette
ggboxplot(df, "dose", "len", color = "dose",
 palette = "Dark2", add = "jitter" )

# Use grey palette
ggboxplot(df, "dose", "len", color = "dose",
palette = "grey" )


# Box plot with multiple groups
# +++++++++++++++++++++
# fill or color box plot by a second group : "supp"
ggboxplot(df, "dose", "len", color = "supp",
 palette = c('#999999','#E69F00') )

# Add dot
ggboxplot(df, "dose", "len", color = "supp",
 palette = c('#999999','#E69F00'), add = "dotplot")

}
\seealso{
\code{\link{ggpar}}
}

