% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdensity.R
\name{ggdensity}
\alias{ggdensity}
\title{Density plot}
\usage{
ggdensity(data, x, y = "..density..", color = "black", fill = NA,
  palette = NULL, size = 1, linetype = "solid", alpha = 0.5,
  add = c("none", "mean", "median"), add.params = list(linetype = "dashed"),
  rug = FALSE, ggtheme = theme_pubr(), ...)
}
\arguments{
\item{data}{a data frame}

\item{x}{variable to be drawn.}

\item{y}{one of "..density.." or "..count..".}

\item{color, fill}{density line color and fill color.}

\item{palette}{the color palette to be used for coloring or filling by
groups. Allowed values include "grey" for grey color palettes; RColorBrewer
palettes e.g. "RdBu", "Blues", ...; or a character vector containing custom
color palette e.g. c("blue", "red").}

\item{size}{change the size of points and outlines.}

\item{linetype}{line type. See \code{\link{show_line_types}}.}

\item{alpha}{numeric value specifying fill color transparency. Value should
be in [0, 1], where 0 is full transparency and 1 is no transparency.}

\item{add}{allowed values are one of "mean" or "median" (for adding mean or
median line, respectively).}

\item{add.params}{parameters (color, size, linetype) for the argument 'add';
e.g.: add.params = list(color = "red").}

\item{rug}{logical value. If TRUE, add marginal rug.}

\item{ggtheme}{function, ggplot2 theme name. Default value is theme_pubr().
Allowed values include ggplot2 official themes: theme_gray(), theme_bw(),
theme_minimal(), theme_classic(), theme_void(), ....}

\item{...}{other arguments to be passed to
\code{\link[ggplot2]{geom_density}} and \code{\link{ggpar}}.}
}
\description{
Create a density plot.
}
\details{
The plot can be easily customized using the function ggpar(). Read
  ?ggpar for changing: \itemize{ \item main title and axis labels: main,
  xlab, ylab \item axis limits: xlim, ylim (e.g.: ylim = c(0, 30)) \item axis
  scales: xscale, yscale (e.g.: yscale = "log2") \item color palettes:
  palette = "Dark2" or palette = c("gray", "blue", "red") \item legend title,
  labels and position: legend = "right" \item plot orientation : orientation
  = c("vertical", "horizontal", "reverse") }
}
\examples{
# Create some data format
set.seed(1234)
wdata = data.frame(
   sex = factor(rep(c("F", "M"), each=200)),
   weight = c(rnorm(200, 55), rnorm(200, 58)))

head(wdata, 4)

# Basic density plot with mean line
ggdensity(wdata, x = "weight", fill = "lightgray",
   add = "mean")

# Change outline and fill colors by groups ("sex")
ggdensity(wdata, x = "weight",
   color = "sex", fill = "sex",
   add = "mean", rug = TRUE)

# Change color palette: custom palette
ggdensity(wdata, x = "weight",
   color = "sex",  add = "mean", rug = TRUE,
   palette = c("#999999", "#E69F00"))

# Use brewer palette
ggdensity(wdata, x = "weight",
   color = "sex", add = "mean", rug = TRUE,
   palette = "Paired")

}
\seealso{
\code{\link{ggpar}}
}

