% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpar.R
\name{ggpar}
\alias{ggpar}
\title{Graphical parameters}
\usage{
ggpar(p, palette = NULL, main = NULL, xlab = NULL, ylab = NULL,
  font.main = NULL, font.x = NULL, font.y = NULL, xlim = NULL,
  ylim = NULL, xscale = c("none", "log2", "log10", "sqrt"),
  yscale = c("none", "log2", "log10", "sqrt"), format.scale = FALSE,
  legend = c("bottom", "top", "left", "right", "none"), legend.title = NULL,
  font.legend = NULL, ticks = TRUE, tickslab = TRUE,
  font.tickslab = NULL, xtickslab.rt = 0, ytickslab.rt = 0,
  xticks.by = NULL, yticks.by = NULL, orientation = c("vertical",
  "horizontal", "reverse"), ...)
}
\arguments{
\item{p}{an object of class ggplot}

\item{palette}{the color palette to be used for coloring or filling by
groups. Allowed values include "grey" for grey color palettes; brewer
palettes e.g. "RdBu", "Blues", ...; or custom color palette e.g. c("blue",
"red").}

\item{main}{plot main title.}

\item{xlab, ylab}{character vector specifying x and y axis labels,
respectively. Use xlab = FALSE and ylab = FALSE to hide xlab and ylab,
respectively.}

\item{font.main, font.x, font.y}{a vector of length 3 indicating respectively
the size (e.g.: 14), the style (e.g.: "plain", "bold", "italic",
"bold.italic") and the color (e.g.: "red") of main title, xlab and ylab,
respectively. For example \emph{font.x = c(14, "bold", "red")}.}

\item{xlim, ylim}{a numeric vector of length 2, specifying  x and y axis limits (minimum
and maximum), respectively. e.g.: ylim = c(0, 50).}

\item{xscale, yscale}{x and y axis scale, respectively. Allowed values are one of c("none", "log2",
"log10", "sqrt"); e.g.: yscale="log2".}

\item{format.scale}{logical value. If TRUE, axis tick mark labels will be
formatted when xscale or yscale = "log2" or "log10".}

\item{legend}{character specifying legend position. Allowed values are one of
c("top", "bottom", "left", "right", "none"). Default is "top" side
position. to remove the legend use legend = "none". Legend position can be
also specified using a numeric vector c(x, y); see details section.}

\item{legend.title}{legend title.}

\item{font.legend}{legend text font style; e.g.: font.legend = c(10, "plain",
"black").}

\item{ticks}{logical value. Default is TRUE. If FALSE, hide axis tick marks.}

\item{tickslab}{logical value. Default is TRUE. If FALSE, hide axis tick
labels.}

\item{font.tickslab}{Font style (size, face, color) for tick labels, e.g.:
c(14, "bold", "red").}

\item{xtickslab.rt, ytickslab.rt}{Rotation angle of x and y axis tick labels,
respectively. Default value is 0.}

\item{xticks.by, yticks.by}{numeric value controlling x and y axis breaks, respectively. For example, if
yticks.by = 5, a tick mark is shown on every 5. Default value is NULL.}

\item{orientation}{change the orientation of the plot. Allowed values are one
of c( "vertical", "horizontal", "reverse"). Partial match is allowed.}

\item{...}{not used}
}
\description{
Graphical parameters
}
\examples{
# Load data
data("ToothGrowth")
df <- ToothGrowth

# Basic box plot
# +++++++++++++++++++++++++++

p <- ggboxplot(df, x = "dose", y = "len")

# Change the plot orientation: horizontal
ggpar(p, orientation = "horiz")


 # Change main title and axis labels
 # ++++++++++++++++++++++++++++

 ggpar(p,
   main = "Plot of length \\n by dose",
   xlab = "Dose (mg)", ylab = "Length")

 # Title font styles: 'plain', 'italic', 'bold', 'bold.italic'
 ggpar(p,
   main = "Length by dose",
   font.main = c(14,"bold.italic", "red"),
   font.x = c(14, "bold", "#2E9FDF"),
   font.y = c(14, "bold", "#E7B800"))

 # Hide axis labels
 ggpar(p, xlab = FALSE, ylab = FALSE)


# Change colors
# ++++++++++++++++++++++

# Change outline colors by groups: dose
 p2 <- ggboxplot(df, "dose", "len", color = "dose")
 p2

# Use custom color palette
ggpar(p2, palette = c('#999999','#E69F00','#56B4E9'))

# Use brewer palette
ggpar(p2, palette = "Dark2" )

# Use grey palette
ggpar(p2, palette = "grey")


# Axis ticks, limits, scales
# +++++++++++++++++++++++++

# Axis ticks labels and rotation
ggpar(p,
 font.tickslab = c(14,"bold", "#993333"),
 xtickslab.rt = 45, ytickslab.rt = 45)
# Hide axis ticks and tick labels
ggpar(p, ticks = FALSE, tickslab = FALSE)

# Axis limits
ggpar(p, ylim = c(0, 50))

# Axis scale
ggpar(p, yscale = "log2", format.scale = TRUE)

# Legends
# ++++++++++++++++++
# Change legend position and title
ggpar(p2,
 legend = "right", legend.title = "Dose (mg)",
 font.legend = c(10, "bold", "red"))

}

