% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscatter.R
\name{ggscatter}
\alias{ggscatter}
\title{Scatter plot}
\usage{
ggscatter(data, x, y, color = "black", fill = "lightgray", palette = NULL,
  shape = 19, size = 2, point = TRUE, rug = FALSE, add = c("none",
  "reg.line", "loess"), add.params = list(), conf.int = FALSE,
  conf.int.level = 0.95, fullrange = FALSE, ellipse = FALSE,
  ellipse.level = 0.95, ellipse.type = "norm", ellipse.alpha = 0.1,
  mean.point = FALSE, mean.point.size = 2 * size, label = NULL,
  label.size = 5, label.select = NULL, repel = FALSE,
  ggtheme = theme_pubr(), ...)
}
\arguments{
\item{data}{a data frame}

\item{x, y}{x and y variables for drawing.}

\item{color, fill}{point colors.}

\item{palette}{the color palette to be used for coloring or filling by
groups. Allowed values include "grey" for grey color palettes; RColorBrewer
palettes e.g. "RdBu", "Blues", ...; or a character vector containing custom
color palette e.g. c("blue", "red").}

\item{shape}{point shape. See \code{\link{show_point_shapes}}.}

\item{size}{change the size of points and outlines.}

\item{point}{logical value. If TRUE, show points.}

\item{rug}{logical value. If TRUE, add marginal rug.}

\item{add}{allowed values are one of "none", "reg.line" (for adding linear
regression line) or "loess" (for adding local regression fitting).}

\item{add.params}{parameters (color, size, linetype) for the argument 'add';
e.g.: add.params = list(color = "red").}

\item{conf.int}{logical value. If TRUE, adds confidence interval.}

\item{conf.int.level}{Level controlling confidence region. Default is 95\%.
Used only when add != "none" and conf.int = TRUE.}

\item{fullrange}{should the fit span the full range of the plot, or just the
data. Used only when add != "none".}

\item{ellipse}{logical value. If TRUE, draws ellipses around points.}

\item{ellipse.level}{the size of the concentration ellipse in normal
probability.}

\item{ellipse.type}{Character specifying frame type. Possible values are
'convex', 'confidence' or types supporeted by
\code{\link[ggplot2]{stat_ellipse}} including one of c("t", "norm",
"euclid").}

\item{ellipse.alpha}{Alpha for ellipse specifying the transparency level of
fill color. Use alpha = 0 for no fill color.}

\item{mean.point}{logical value. If TRUE, group mean points are added to the
plot.}

\item{mean.point.size}{numeric value specifying the size of mean points.}

\item{label}{the name of the column containing point labels.}

\item{label.size}{text size for point labels.}

\item{label.select}{character vector specifying some labels to show.}

\item{repel}{a logical value, whether to use ggrepel to avoid overplotting text
labels or not.}

\item{ggtheme}{function, ggplot2 theme name. Default value is theme_pubr().
Allowed values include ggplot2 official themes: theme_gray(), theme_bw(),
theme_minimal(), theme_classic(), theme_void(), ....}

\item{...}{other arguments to be passed to \code{\link[ggplot2]{geom_point}}
and \code{\link{ggpar}}.}
}
\description{
Create a scatter plot.
}
\details{
The plot can be easily customized using the function ggpar(). Read
  ?ggpar for changing: \itemize{ \item main title and axis labels: main,
  xlab, ylab \item axis limits: xlim, ylim (e.g.: ylim = c(0, 30)) \item axis
  scales: xscale, yscale (e.g.: yscale = "log2") \item color palettes:
  palette = "Dark2" or palette = c("gray", "blue", "red") \item legend title,
  labels and position: legend = "right" \item plot orientation : orientation
  = c("vertical", "horizontal", "reverse") }
}
\examples{
# Load data
data("mtcars")
df <- mtcars
df$cyl <- as.factor(df$cyl)
head(df[, c("wt", "mpg", "cyl")], 3)

# Basic plot
# +++++++++++++++++++++++++++
ggscatter(df, x = "wt", y = "mpg",
   color = "black", shape = 21, size = 3, # point color, shape and size
   add = "reg.line", add.params = list(color = "blue", fill = "lightgray"),
   conf.int = TRUE)

# loess method: local regression fitting
ggscatter(df, x = "wt", y = "mpg",
   add = "loess", conf.int = TRUE)


# Control point size by continuous variable values ("qsec")
ggscatter(df, x = "wt", y = "mpg",
   color = "#00AFBB", size = "qsec")

# Add marginal rug
# +++++++++++++++++++++++++++++
ggscatter(df, x = "wt", y = "mpg",
   color = "cyl", rug = TRUE)


# Change colors
# +++++++++++++++++++++++++++
# Use custom color palette
ggscatter(df, x = "wt", y = "mpg", color = "cyl",
   palette = c('#999999','#E69F00','#56B4E9') )

# Use brewer palette
ggscatter(df, x = "wt", y = "mpg", color = "cyl",
   palette = "Dark2" )

# Use grey palette
ggscatter(df, x = "wt", y = "mpg", color = "cyl",
   palette = "grey" )


# Add group ellipses and mean points
# +++++++++++++++++++

ggscatter(df, x = "wt", y = "mpg",
   color = "cyl", shape = "cyl",
   palette = c('#999999','#E69F00','#56B4E9'),
   ellipse = TRUE, mean.point = TRUE)


# Textual annotation



}
\seealso{
\code{\link{ggpar}}
}

