% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstripchart.R
\name{ggstripchart}
\alias{ggstripchart}
\title{Stripcharts}
\usage{
ggstripchart(data, x, y, color = "black", fill = "white", palette = NULL,
  shape = 19, select = NULL, order = NULL, add = c("none", "boxplot",
  "violin", "pointrange", "crossbar"), add.params = list(),
  position = position_jitter(0.4), ...)
}
\arguments{
\item{data}{a data frame}

\item{x, y}{x and y variables for drawing.}

\item{color, fill}{outline and fill colors.}

\item{palette}{the color palette to be used for coloring or filling by
groups. Allowed values include "grey" for grey color palettes; RColorBrewer
palettes e.g. "RdBu", "Blues", ...; or a character vector containing custom
color palette e.g. c("blue", "red").}

\item{shape}{point shape}

\item{select}{character vector specifying which items to display.}

\item{order}{character vector specifying the order of items.}

\item{add}{character vector for adding another plot element (e.g.: dot plot
or error bars). Allowed values are one or the combination of: "none", "dotplot", "jitter",
"mean", "mean_se", "mean_sd", "mean_ci", "mean_range", "median",
"median_iqr", "median_mad", "median_range".}

\item{add.params}{parameters (color, shape, size, fill, linetype) for the
argument 'add'; e.g.: add.params = list(color = "red").}

\item{position}{position adjustment, either as a string, or the result of a
call to a position adjustment function. Used to adjust position
for multiple groups.}

\item{...}{other arguments to be passed to geom_jitter.}
}
\description{
Stripcharts
}
\details{
The plot can be easily customized using the function ggpar(). Read
  ?ggpar for changing: \itemize{ \item main title and axis labels: main,
  xlab, ylab \item axis limits: xlim, ylim (e.g.: ylim = c(0, 30)) \item axis
  scales: xscale, yscale (e.g.: yscale = "log2") \item color palettes:
  palette = "Dark2" or palette = c("gray", "blue", "red") \item legend title,
  labels and position: legend = "right" \item plot orientation : orientation
  = c("vertical", "horizontal", "reverse") }
}
\examples{
# Load data
data("ToothGrowth")
df <- ToothGrowth

# Basic box plot
# +++++++++++++++++++++++++++
ggstripchart(df, x = "dose", y = "len")
# Change point shapes and size
ggstripchart(df, "dose", "len", shape = 17, size = 2)


# Select and order items
# ++++++++++++++++++++++++++++++
# Select which items to display: "0.5" and "2"
ggstripchart(df, "dose", "len",
   select = c("0.5", "2"))
# Change the default order of items
ggstripchart(df, "dose", "len",
   order = c("2", "1", "0.5"))

# Add summary statistics
# ++++++++++++++++++++++++++
# Add box plot
ggstripchart(df, x = "dose", y = "len",
add = "boxplot")
# Add violin
ggstripchart(df, x = "dose", y = "len",
add = "violin")
# Add pointrange
ggstripchart(df, x = "dose", y = "len",
add = "pointrange")
# Add crossbar
ggstripchart(df, x = "dose", y = "len",
add = "crossbar", add.params = list(width = 0.5))
# Add violin + pointrange
ggstripchart(df, x = "dose", y = "len", color = "dose",
add = c("violin", "pointrange"))



# Change colors
# +++++++++++++++++++++++++++
# Change colors
 ggstripchart(df, "dose", "len", color = "#2E9FDF",
  add = "pointrange", add.params = list(color = "red"))

# Change colors by groups: dose
 ggstripchart(df, "dose", "len", color = "dose",
   add = "boxplot")

# Use custom color palette
 ggstripchart(df, "dose", "len", color = "dose",
  palette = c('#999999','#E69F00','#56B4E9') )

# Use brewer palette
ggstripchart(df, "dose", "len", color = "dose",
 palette = "Dark2")

# Use grey palette
ggstripchart(df, "dose", "len", color = "dose",
palette = "grey" )


# Plot with multiple groups
# +++++++++++++++++++++
# Change shape and color by a second group : "supp"
ggstripchart(df, "dose", "len", color = "supp",
  shape = "supp" )

# Add boxplot
ggstripchart(df, "dose", "len", color = "supp", shape = "supp",
 add = "boxplot", add.params = list(color = "black") )

# Change point position
ggstripchart(df, "dose", "len", color = "supp",
  shape = "supp", add = "boxplot", position = position_dodge(0.8) )


}
\seealso{
\code{\link{ggpar}}
}

