% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviolin.R
\name{ggviolin}
\alias{ggviolin}
\title{Violin plot}
\usage{
ggviolin(data, x, y, color = "black", fill = "white", palette = NULL,
  linetype = "solid", trim = FALSE, size = 1, width = 1,
  draw_quantiles = NULL, select = NULL, order = NULL, add = c("none",
  "boxplot", "dotplot", "jitter", "pointrange", "crossbar"),
  add.params = list(), ...)
}
\arguments{
\item{data}{a data frame}

\item{x, y}{x and y variables for drawing.}

\item{color, fill}{outline and fill colors.}

\item{palette}{the color palette to be used for coloring or filling by
groups. Allowed values include "grey" for grey color palettes; RColorBrewer
palettes e.g. "RdBu", "Blues", ...; or a character vector containing custom
color palette e.g. c("blue", "red").}

\item{linetype}{line types.}

\item{trim}{If \code{TRUE} (default), trim the tails of the violins
to the range of the data. If \code{FALSE}, don't trim the tails.}

\item{size}{line size.}

\item{width}{violin width.}

\item{draw_quantiles}{If \code{not(NULL)} (default), draw horizontal lines
at the given quantiles of the density estimate.}

\item{select}{character vector specifying which items to display.}

\item{order}{character vector specifying the order of items.}

\item{add}{character vector for adding another plot element (e.g.: dot plot
or error bars). Allowed values are one or the combination of: "none", "dotplot", "jitter",
"mean", "mean_se", "mean_sd", "mean_ci", "mean_range", "median",
"median_iqr", "median_mad", "median_range".}

\item{add.params}{parameters (color, shape, size, fill, linetype) for the
argument 'add'; e.g.: add.params = list(color = "red").}

\item{...}{other arguments to be passed to geom_violin.}
}
\description{
Violin plot
}
\details{
The plot can be easily customized using the function ggpar(). Read
  ?ggpar for changing: \itemize{ \item main title and axis labels: main,
  xlab, ylab \item axis limits: xlim, ylim (e.g.: ylim = c(0, 30)) \item axis
  scales: xscale, yscale (e.g.: yscale = "log2") \item color palettes:
  palette = "Dark2" or palette = c("gray", "blue", "red") \item legend title,
  labels and position: legend = "right" \item plot orientation : orientation
  = c("vertical", "horizontal", "reverse") }
}
\examples{
# Load data
data("ToothGrowth")
df <- ToothGrowth

# Basic plot
# +++++++++++++++++++++++++++
ggviolin(df, x = "dose", y = "len")
# Change width
ggviolin(df, "dose", "len", width = 0.2)
# Change the plot orientation: horizontal
ggviolin(df, "dose", "len", orientation = "horiz")

# Add summary statistics
# ++++++++++++++++++++++++++
# Draw quantiles
ggviolin(df, "dose", "len", add = "none",
   draw_quantiles = 0.5)

# Add box plot
ggviolin(df, x = "dose", y = "len",
 add = "boxplot")

ggviolin(df, x = "dose", y = "len",
 add = "dotplot")

# Add jitter
ggviolin(df, x = "dose", y = "len",
add = "jitter")

# Change shape by groups: "dose"
ggviolin(df, x = "dose", y = "len",
add = "jitter", shape = "dose")

# Add pointrange
ggviolin(df, x = "dose", y = "len",
 add = "pointrange")

# Add crossbar
ggviolin(df, x = "dose", y = "len",
 add = "crossbar")

# Combine jitter + pointrange
ggviolin(df, x = "dose", y = "len",
 add = c("jitter", "pointrange"))



# Select and order items
# ++++++++++++++++++++++++++++++
# Select which items to display: "0.5" and "2"
ggviolin(df, "dose", "len",
   select = c("0.5", "2"))
# Change the default order of items
ggviolin(df, "dose", "len",
   order = c("2", "1", "0.5"))


# Change colors
# +++++++++++++++++++++++++++
# Change outline and fill colors
# Change line size
ggviolin(df, "dose", "len",
   color = "#2E9FDF", fill = "#6D9EC1", size = 1.2)

# Change outline colors by groups: dose
ggviolin(df, "dose", "len", color = "dose")

# Change fill color by groups: dose
ggviolin(df, "dose", "len", fill = "dose")

# Use custom color palette
ggviolin(df, "dose", "len", color = "dose",
  palette = c('#999999','#E69F00','#56B4E9') )

# Use brewer palette
ggviolin(df, "dose", "len", color = "dose",
 palette = "Dark2", add = "boxplot" )

# Use grey palette
ggviolin(df, "dose", "len", color = "dose",
palette = "grey" )


# Plot with multiple groups
# +++++++++++++++++++++
# fill or color box plot by a second group : "supp"
ggviolin(df, "dose", "len", color = "supp",
 palette = c('#999999','#E69F00') )

# Add dot
ggviolin(df, "dose", "len", color = "supp",
 palette = c('#999999','#E69F00'), add = "dotplot")

}
\seealso{
\code{\link{ggpar}}
}

