% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{mutate}
\alias{group_by}
\alias{geom_signif}
\alias{get_summary_stats}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}}

  \item{ggsignif}{\code{\link[ggsignif]{geom_signif}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{rstatix}{\code{\link[rstatix]{get_summary_stats}}}
}}

