% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate a ggplot Horizontally}
\usage{
rotate(...)
}
\arguments{
\item{...}{other arguments to pass to \code{\link[ggplot2]{coord_flip}}.}
}
\description{
Rotate a ggplot to create horizontal plots. Wrapper around
 \code{\link[ggplot2]{coord_flip}}.
}
\examples{
# Load data
data("ToothGrowth")
df <- ToothGrowth

# Basic plot
p <- ggboxplot(df, x = "dose", y = "len",
   color = "dose", palette = "jco")
p
# Create horizontal plots
p + rotate()
}
