% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_chull.R
\name{stat_chull}
\alias{stat_chull}
\title{Plot convex hull of a set of points}
\usage{
stat_chull(mapping = NULL, data = NULL, geom = "path",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
Plot convex hull of a set of points.
}
\examples{
# Load data
data("mtcars")
df <- mtcars
df$cyl <- as.factor(df$cyl)

# scatter plot with convex hull
ggscatter(df, x = "wt", y = "mpg", color = "cyl")+
 stat_chull(aes(color = cyl))

ggscatter(df, x = "wt", y = "mpg", color = "cyl")+
 stat_chull(aes(color = cyl, fill = cyl), alpha = 0.1, geom = "polygon")

}
\seealso{
\code{\link{ggpar}}, \code{\link{ggscatter}}
}
