% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_compare_means.R
\name{stat_compare_means}
\alias{stat_compare_means}
\title{Add Mean Comparison P-values to a ggplot}
\usage{
stat_compare_means(mapping = NULL, data = NULL, method = NULL,
  paired = FALSE, label.sep = ", ", label.x.npc = "left",
  label.y.npc = "top", label.x = NULL, label.y = NULL, geom = "text",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{method}{a character string indicating which method to be used for
comparing means.}

\item{paired}{a logical indicating whether you want a paired test. Used only
in \code{\link[stats]{t.test}} and in \link[stats]{wilcox.test}.}

\item{label.sep}{a character string to separate the terms. Default is ", ",
to separate the correlation coefficient and the p.value.}

\item{label.x.npc, label.y.npc}{can be \code{numeric} or \code{character}
  vector of the same length as the number of groups and/or panels. If too
  short they will be recycled. \itemize{ \item If \code{numeric}, value
  should be between 0 and 1. Coordinates to be used for positioning the
  label, expressed in "normalized parent coordinates". \item If
  \code{character}, allowed values include: i) one of c('right', 'left',
  'center', 'centre', 'middle') for x-axis; ii) and one of c( 'bottom',
  'top', 'center', 'centre', 'middle') for y-axis.}

  If too short they will be recycled.}

\item{label.x, label.y}{\code{numeric} Coordinates (in data units) to be used
for absolute positioning of the label. If too short they will be recycled.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If FALSE (the default), removes missing values with a warning.
If TRUE silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments to pass to \code{\link[ggplot2]{geom_text}} or
\code{\link[ggplot2]{geom_label}}.}
}
\description{
Add mean comprison p-values to a ggplot, such as box blots, dot
  plots and stripcharts.
}
\examples{
# Load data
data("mtcars")
df <- mtcars
df$cyl <- as.factor(df$cyl)

# Scatter plot with correlation coefficient
#:::::::::::::::::::::::::::::::::::::::::::::::::
sp <- ggscatter(df, x = "wt", y = "mpg",
   add = "reg.line",  # Add regressin line
   add.params = list(color = "blue", fill = "lightgray"), # Customize reg. line
   conf.int = TRUE # Add confidence interval
   )
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = 3, label.y = 30)

# Color by groups and facet
#::::::::::::::::::::::::::::::::::::::::::::::::::::
sp <- ggscatter(df, x = "wt", y = "mpg",
   color = "cyl", palette = "jco",
   add = "reg.line", conf.int = TRUE)
sp + stat_cor(aes(color = cyl), label.x = 3)

}
\seealso{
\code{\link{ggscatter}}
}
