% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_conf_ellipse.R
\name{stat_conf_ellipse}
\alias{stat_conf_ellipse}
\title{Plot confidence ellipses.}
\usage{
stat_conf_ellipse(mapping = NULL, data = NULL, geom = "path",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, level = 0.95, npoint = 100, bary = TRUE, ...)
}
\arguments{
\item{level}{confidence level used to construct the ellipses. By
default, 0.95.}

\item{npoint}{number of points used to draw the ellipses.}

\item{bary}{logical value. If TRUE, the coordinates of the ellipse around the
barycentre of individuals are calculated.}
}
\description{
Plot confidence ellipses around barycenters. The method for
  computing confidence ellipses has been modified from FactoMineR::coord.ellipse.
}
\examples{
# Load data
data("mtcars")
df <- mtcars
df$cyl <- as.factor(df$cyl)

# scatter plot with confidence ellipses
ggscatter(df, x = "wt", y = "mpg", color = "cyl")+
 stat_conf_ellipse(aes(color = cyl))

ggscatter(df, x = "wt", y = "mpg", color = "cyl")+
 stat_conf_ellipse(aes(color = cyl, fill = cyl), alpha = 0.1, geom = "polygon")

}
\seealso{
\code{\link{stat_conf_ellipse}}
}
