% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mean.R
\name{stat_mean}
\alias{stat_mean}
\title{Draw group mean points}
\usage{
stat_mean(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{na.rm}{If FALSE (the default), removes missing values with a warning.
If TRUE silently removes missing values.}

\item{...}{other arguments to pass to \code{\link[ggplot2]{geom_point}}.}
}
\description{
Draw the mean point of each group.
}
\examples{
# Load data
data("mtcars")
df <- mtcars
df$cyl <- as.factor(df$cyl)

# Scatter plot with ellipses and group mean points
ggscatter(df, x = "wt", y = "mpg",
   color = "cyl", shape = "cyl", ellipse = TRUE)+
 stat_mean(aes(color = cyl, shape = cyl), size = 4)

}
\seealso{
\code{\link{stat_conf_ellipse}}, \code{\link{stat_chull}} and
  \code{\link{ggscatter}}
}
