% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_pvalue_manual.R
\name{stat_pvalue_manual}
\alias{stat_pvalue_manual}
\title{Add Manually P-values to a ggplot}
\usage{
stat_pvalue_manual(data, label = "p", y.position = "y.position",
  xmin = "group1", xmax = "group2", tip.length = 0.03, ...)
}
\arguments{
\item{data}{a data frame containing statitistical test results. The expected
default format should contain the following columns: \code{group1 | group2 |
p | y.position | etc}. \code{group1} and \code{group2} are the groups that
have been compared. \code{p} is the resulting p-value. \code{y.position} is
the y coordinates of the p-values in the plot.}

\item{label}{the column containing the label (e.g.: label = "p"). Default
value is "p", for p-value.}

\item{y.position}{\code{numeric} column containing the coordinates (in data
units) to be used for absolute positioning of the label. Default value is
"y.position".}

\item{xmin}{column containing the position of the left sides of the brackets.
Default value is "group1".}

\item{xmax}{(optional) column containing the position of the right sides of
the brackets. Default value is "group2". If NULL, the p-values are plotted
as a simple text.}

\item{tip.length}{numeric vector with the fraction of total height that the
bar goes down to indicate the precise column. Default is 0.03.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Add manually p-values to a ggplot, such as box blots, dot plots
 and stripcharts.
}
\examples{

require("dplyr")

#:::::::::::::::::::::::::::::::::::
# Simple box plot
#:::::::::::::::::::::::::::::::::::
# Pairwise t-test between groups
stat.test <- compare_means(
  len ~ dose, data = ToothGrowth,
  method = "t.test"
) \%>\%
  mutate(y.position = c(29, 35, 39))
stat.test
# Create a box plot and add the p-value
ggboxplot(ToothGrowth, x = "dose", y = "len") +
  stat_pvalue_manual(stat.test, label = "p.adj")
}
\seealso{
\code{\link{stat_compare_means}}
}
