% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_stars.R
\name{stat_stars}
\alias{stat_stars}
\title{Add Stars to a Scatter Plot}
\usage{
stat_stars(mapping = NULL, data = NULL, geom = "segment",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{na.rm}{If FALSE (the default), removes missing values with a warning.
If TRUE silently removes missing values.}

\item{...}{other arguments to pass to \code{\link[ggplot2]{geom_segment}}.}
}
\description{
Create a star plot by drawing segments from group centroid to each points.
}
\examples{
# Load data
data("mtcars")
df <- mtcars
df$cyl <- as.factor(df$cyl)

# Scatter plot with ellipses and group mean points
ggscatter(df, x = "wt", y = "mpg",
   color = "cyl", shape = "cyl",
   mean.point = TRUE, ellipse = TRUE)+
 stat_stars(aes(color = cyl))

}
\seealso{
\code{\link{ggscatter}}
}
